[Feat] Elastic Search 전환 완료

+++ backend_admin/src/main/java/com/example/backend_admin/elastic/entity/OrdersDocument.java
@@ -24,11 +24,11 @@ public class OrdersDocument {
     @Field(type = FieldType.Integer)
     private int productIdx;
 
-    @Field(type = FieldType.Double)
-    private double category;
+    @Field(type = FieldType.Integer)
+    private int category;
 
-    @Field(type = FieldType.Double)
-    private double price;
+    @Field(type = FieldType.Integer)
+    private int price;
 
     @Field(type = FieldType.Text)
     private String impUid;
